requirements.txt
python-telegram-bot
openai
pydub
